SERVICES-1554: update user energy to use GetOrSet decorator
- removed user.energy.getter.service.ts
- updated dependencies to use user energy compute service
